PlasmaProperties.jl

Documentation for PlasmaProperties.jl

PlasmaProperties — Module

PlasmaProperties, a simple plasma calculator module in Julia

The module contains basic functions to compute common plasma properties, such as characteristic frequencies and lengths.

It builds on Unitful, which provides a simple unit system for physical quantities.

Module Index

PlasmaProperties.f_ce

PlasmaProperties.f_pe

PlasmaProperties.λ_D

Detailed API

PlasmaProperties.f_ce — Method

f_ce(B)

Compute the cyclotron frequency of electrons in MHz. If B does not have Unitful units, it is assumed to be in Tesla.

PlasmaProperties.f_pe — Method

f_pe(ne)

Compute the plasma frequency of electrons in MHz. If ne does not have Unitful units, it is assumed to be in m^-3.

PlasmaProperties.λ_D — Method

λ_D(ne,Te)

Compute the Debye length of the plasm in m. If ne does not have Unitful units, it is assumed to be in m^-3. If Te does not have Unitful units, it is assumed to be in eV.