PlasmaProperties.jl

Documentation for PlasmaProperties.jl

PlasmaProperties — Module

PlasmaProperties, a simple plasma calculator module in Julia

The module contains basic functions to compute common plasma properties, such as characteristic frequencies and lengths.

It builds on Unitful, which provides a simple unit system for physical quantities.

Module Index

PlasmaProperties.Plasma

PlasmaProperties.check_or_assume

PlasmaProperties.f_ce

PlasmaProperties.f_ci

PlasmaProperties.f_pe

PlasmaProperties.f_pi

PlasmaProperties.force_energy_units!

PlasmaProperties.rLe

PlasmaProperties.λ_De

Detailed API

PlasmaProperties.Plasma — Type

P = Plasma()

Mutable type holding plasma properties like density, temperature, etc. The show() method prints a table of the properties and derived quantities.

A name can be specified to quickly select the ion mass and charge. If units are not given, the default is mi in kg, n in 1/m^3, Te in eV, B in Gauss.

PlasmaProperties.check_or_assume — Method

checkorassume(value,units)

If value is not a Unitful quantity, assume it is in the given units.

PlasmaProperties.f_ce — Method

f_ce(B)
f_ce(P::Plasma)

Compute the (unsigned) cyclotron frequency of electrons in MHz. If B does not have Unitful units, it is assumed to be in Tesla.

PlasmaProperties.f_ci — Method

f_ci(mi,Z,B)
f_ci(P::Plasma)

Compute the cyclotron frequency of ions of mass mi in MHz. If B does not have Unitful units, it is assumed to be in Tesla.

PlasmaProperties.f_pe — Method

f_pe(ne)
f_pe(P::Plasma)

Compute the (unsigned) plasma frequency of electrons in MHz. If ne does not have Unitful units, it is assumed to be in m^-3.

PlasmaProperties.f_pi — Method

f_pi(mi,Z,ni)
f_pi(P::Plasma)

Compute the plasma frequency of ions of mass mi in MHz. If ni does not have Unitful units, it is assumed to be in m^-3.

PlasmaProperties.force_energy_units! — Method

force_energy_units!(value)

If value is in temperature units, convert it to energy units.

PlasmaProperties.rLe — Method

rLe(B,vperpe)

Compute the electron gyro radius in m. If B does not have Unitful units, it is assumed to be in Tesla. If vperpe does not have Unitful units, it is assumed to be in m/s.

PlasmaProperties.λ_De — Method

λ_De(ne,Te)
λ_De(P::Plasma)

Compute the Debye length of the plasm in m. If ne does not have Unitful units, it is assumed to be in m^-3. If Te does not have Unitful units, it is assumed to be in eV.